refactor(problem): 난이도, 답안 유형정보 core 모듈로 이동
