+++ domain/mathrank-problem-core-domain/src/main/java/kr/co/mathrank/domain/problem/core/Difficulty.java
@@ -1,4 +1,4 @@
-package kr.co.mathrank.domain.problem.entity;
+package kr.co.mathrank.domain.problem.core;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
