+++ domain/mathrank-problem-domain/build.gradle
@@ -1,3 +1,7 @@
+plugins {
+    id 'java-library'
+}
+
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
@@ -8,6 +12,7 @@ dependencies {
 
     implementation project(':common:mathrank-snowflake')
     implementation project(':common:mathrank-exception')
+    api project(':domain:mathrank-problem-core-domain')
 
     implementation 'com.querydsl:querydsl-jpa:5.0.0:jakarta'
     annotationProcessor "com.querydsl:querydsl-apt:${dependencyManagement.importedProperties['querydsl.version']}:jakarta"
