+++ domain/mathrank-problem-domain/src/main/java/kr/co/mathrank/domain/problem/dto/ProblemRegisterCommand.java
@@ -5,8 +5,8 @@ import java.util.Set;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
-import kr.co.mathrank.domain.problem.entity.AnswerType;
-import kr.co.mathrank.domain.problem.entity.Difficulty;
+import kr.co.mathrank.domain.problem.core.AnswerType;
+import kr.co.mathrank.domain.problem.core.Difficulty;
 
 public record ProblemRegisterCommand(
 	@NotNull
