+++ domain/mathrank-problem-domain/src/main/java/kr/co/mathrank/domain/problem/dto/ProblemUpdateCommand.java
@@ -4,8 +4,8 @@ import java.util.Set;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
-import kr.co.mathrank.domain.problem.entity.AnswerType;
-import kr.co.mathrank.domain.problem.entity.Difficulty;
+import kr.co.mathrank.domain.problem.core.AnswerType;
+import kr.co.mathrank.domain.problem.core.Difficulty;
 
 public record ProblemUpdateCommand(
 	@NotNull
