+++ domain/mathrank-problem-domain/src/test/java/kr/co/mathrank/domain/problem/entity/ProblemTest.java
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 
+import kr.co.mathrank.domain.problem.core.AnswerType;
+import kr.co.mathrank.domain.problem.core.Difficulty;
 import kr.co.mathrank.domain.problem.repository.ProblemRepository;
 
 @SpringBootTest
